Make 0xFF converter run again
Ignore several new abuses of instant_messenger_nick.
Allow mentoring/mentored persons to be a dupe (look up in person_dupes).
Add option to selectively ignore sub-IPs of spidered devices.
Automatically ignore IP-clash if the primary IP of a device is re-used
as the secondary IP of another device (secondary is ignored). This
already triggered a warning (and still does).